# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ailurus Backend API",
    "version": "2.0.0",
    "description": "This is the API documentation for the Ailurus CTF platform backend."
  },
  "servers": [
    {
      "url": "/api/v2"
    }
  ],
...
  "paths": {
...
    "/admin/contests/reset-game/": {
        "post": {
            "tags": ["admin"],
            "summary": "Reset game contest data (admin)",
            "requestBody": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "confirm": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Game contest data reset"
                }
            }
        }
    },
    "/admin/contests/rotate-flag/": {
        "post": {
            "tags": ["admin"],
            "summary": "Force rotate flag for current tick and round (admin)",
            "requestBody": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "confirm": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Flag rotation task created"
                }
            }
        }
    },
    "/admin/contests/dry-run/": {
        "post": {
            "tags": ["admin"],
            "summary": "Dry run the contest (admin)",
            "requestBody": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "teams": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                },
                                "challenges": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Dry run triggered"
                }
            }
        }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}